Revert "Rename package to functional to distinguish different types of tests"
This reverts commit 237ef35f
