+++ gateway-service/src/test/java/org/zowe/apiml/acceptance/common/AcceptanceTestWithBasePath.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package org.zowe.apiml.functional.common;
+package org.zowe.apiml.acceptance.common;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.web.server.LocalServerPort;
