+++ gateway-service/src/test/java/org/zowe/apiml/acceptance/common/Service.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package org.zowe.apiml.functional.common;
+package org.zowe.apiml.acceptance.common;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
