+++ gateway-service/src/test/java/org/zowe/apiml/acceptance/netflix/ApimlRouteLocatorStub.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package org.zowe.apiml.functional.netflix;
+package org.zowe.apiml.acceptance.netflix;
 
 import org.springframework.cloud.client.discovery.DiscoveryClient;
 import org.springframework.cloud.netflix.zuul.filters.ZuulProperties;
